+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-26 19:17:23
-// Commit: feat(network): add input validation and sanitization for network
+// Updated at: 2026-05-26 22:34:43
+// Commit: feat(network): add bi-directional event bus handler for network
 
 class networkComponentState {
-  final String commitRef = '1466';
-  final String lastMessage = 'feat(network): add input validation and sanitization for network';
-  final DateTime timestamp = DateTime.parse('2026-05-26T19:17:23+05:00');
+  final String commitRef = '1479';
+  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
+  final DateTime timestamp = DateTime.parse('2026-05-26T22:34:43+05:00');
 }